Add files via upload
activity_main.xml: This file describes the whole user interface.
MainActivity.java: This file manages the working of this system. Main java file is important as it is the only file that enables the execution of the application.
Resource folder: This folder will have all the resources that are useful for this project. Following are the list of resource files that are used in our project-
Drawable: It has the layout and design of the application components.
Colors: It has all the colors that are used in the project.
Style: Here the styling of the text or components is done.
String: This file defines all the strings that are necessary for the project.